+++ R/utils-containers.R
@@ -38,7 +38,7 @@ is_singularity_available <- function() {
 }
 
 singularity_cmd <- function(...,
-                            verbose = TRUE,
+                            verbose = FALSE,
                             stdout = "|") {
   singularity_bin_path <- is_singularity_available()
   px_res <- processx::run(